{ipn/ipnlocal, taildrop}: move put logic to taildrop
Cleaning up taildrop logic for sending files.

Updates tailscale/corp#14772
